Various auto changes
currently 2 auto modes, ramparts and the moat, both working really well,
we added in light sensors to tell when we are flat on that other side
again. We will need to add a way to handle different positions the robot
may be in, currently the auto mode is determined from parameters
eventually this will be changed to use a switch on the robot.